// src/types/notification.enum.ts
export enum NotificationType {
  FRIEND_REQUEST = 'friend_request',
  FRIEND_ACCEPTED = 'friend_accepted',
  CHALLENGE_INVITE = 'challenge_invite',
  CHALLENGE_START = 'challenge_start',
  CHALLENGE_END = 'challenge_end',
  CHALLENGE_REMINDER = 'challenge_reminder',
  MISSION_CREATED = 'mission_created',
  MISSION_REMINDER = 'mission_reminder',
  NEW_MESSAGE = 'new_message',
  POST_LIKE = 'post_like',
  POST_COMMENT = 'post_comment',
  MENTION = 'mention',
}